refactor: Extract creation of the store
- Easier testing
- Dry code
