+++ src/targets/browser/index.jsx
@@ -1,4 +1,3 @@
-/* global __DEVELOPMENT__ */
 /* global cozy */
 
 import 'cozy-ui/transpiled/react/stylesheet.css'
@@ -7,33 +6,15 @@ import 'styles/index.styl'
 import React from 'react'
 import { render } from 'react-dom'
 import { Provider, connect } from 'react-redux'
-import { compose, createStore, applyMiddleware } from 'redux'
-import thunkMiddleware from 'redux-thunk'
-import { createLogger } from 'redux-logger'
 import { CozyProvider } from 'cozy-client'
 
 import I18n from 'cozy-ui/transpiled/react/I18n'
 import PiwikHashRouter from 'lib/PiwikHashRouter'
 
-import appReducer from 'reducers'
-
 import App from 'components/App'
 import cozyClient from 'lib/client'
 
-const loggerMiddleware = createLogger()
-
-// Enable Redux dev tools
-const composeEnhancers =
-  (__DEVELOPMENT__ && window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__) || compose
-
-const middlewares = [thunkMiddleware]
-
-if (__DEVELOPMENT__) middlewares.push(loggerMiddleware)
-
-const store = createStore(
-  appReducer,
-  composeEnhancers(applyMiddleware(...middlewares))
-)
+import createStore from '../../store'
 
 const EnhancedI18n = connect(state => {
   const { lang } = state.ui
@@ -56,6 +37,7 @@ document.addEventListener('DOMContentLoaded', () => {
     token: data.cozyToken
   })
 
+  const store = createStore()
   cozyClient.setStore(store)
 
   cozy.bar.init({
